requirements.txt
django
celery==4.0.2
django_redis
mongoengine
psycopg2
django-jet
acora
django-redis-cache
tweepy
tqdm
arabic_reshaper
python-bidi
jdatetime
wordcloud
django-celery-beat==1.0.1